+++ src/main/java/com/heartsave/todaktodak_api/auth/dto/request/LoginIdCheckRequest.java
@@ -1,4 +1,4 @@
-package com.heartsave.todaktodak_api.auth.dto;
+package com.heartsave.todaktodak_api.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
