+++ src/main/java/com/heartsave/todaktodak_api/auth/dto/request/LoginRequest.java
@@ -1,3 +1,3 @@
-package com.heartsave.todaktodak_api.auth.dto;
+package com.heartsave.todaktodak_api.auth.dto.request;
 
 public record LoginRequest(String loginId, String password) {}
